Separating out link tests from service catalog dependent tests
